refactor: move implicit lib to workspace libs
