+++ .vscode/extensions.json
@@ -3,6 +3,7 @@
     "nrwl.angular-console",
     "angular.ng-template",
     "dbaeumer.vscode-eslint",
-    "esbenp.prettier-vscode"
+    "esbenp.prettier-vscode",
+    "firsttris.vscode-jest-runner"
   ]
 }
