+++ angular.json
@@ -1,6 +1,9 @@
 {
   "version": 2,
   "projects": {
-    "bossGtPatchEditor": "apps/bossGtPatchEditor"
+    "bossGtPatchEditor": "apps/bossGtPatchEditor",
+    "shared-api": "libs/shared/api",
+    "shared-effects": "libs/shared/effects",
+    "utils-generators": "libs/utils/generators"
   }
 }
