+++ apps/bossGtPatchEditor/src/app/modules/app/components/demo/demo.component.html
@@ -1,1 +1,2 @@
-<bgt-pe-block-container *ngFor="let block of blocks" [effectBlock]="block" [switchable]="block.switchable" [blockTitle]="block.title" [theme]="block.theme"></bgt-pe-block-container>
+<bgt-pe-block-container *ngFor="let block of blocks" [effectBlock]="block" [switchable]="block.switchable"
+  [blockTitle]="block.title"></bgt-pe-block-container>
