--- apps/bossGtPatchEditor/src/app/modules/data/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from './interfaces.block';
