--- apps/bossGtPatchEditor/src/app/modules/data/types/index.ts
@@ -1,1 +1,0 @@
-export * from './types.block';
