--- apps/bossGtPatchEditor/src/app/modules/data/values/index.ts
@@ -1,1 +1,0 @@
-export * from './blocks';
