--- apps/bossGtPatchEditor/src/app/modules/utils/generators/index.ts
@@ -1,1 +1,0 @@
-export * from './param.generator';
