+++ apps/bossGtPatchEditor/tsconfig.editor.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["**/*.ts"],
+  "compilerOptions": {
+    "types": ["jest", "node"]
+  },
+  "exclude": ["jest.config.ts"]
+}
