+++ libs/shared/api/README.md
@@ -1,0 +1,11 @@
+# shared-api
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-api` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Running lint
+
+Run `nx lint shared-api` to execute the lint via [ESLint](https://eslint.org/).
