+++ libs/shared/api/src/index.ts
@@ -1,0 +1,6 @@
+// interfaces
+export * from './lib/interfaces/effect-block';
+export * from './lib/interfaces/effect-parameter';
+export * from './lib/interfaces/parameter-value';
+// types
+export * from './lib/types/block-type';
