+++ libs/shared/api/src/lib/interfaces/effect-parameter.ts
@@ -1,0 +1,7 @@
+import { ParameterValue } from './parameter-value';
+
+export interface EffectParameter {
+  name: string;
+  values: ParameterValue[];
+  defaultValueIndex?: number;
+}
