+++ libs/shared/api/src/lib/interfaces/parameter-value.ts
@@ -1,0 +1,4 @@
+export interface ParameterValue {
+  display: string;
+  value: number;
+}
