+++ libs/utils/generators/README.md
@@ -1,0 +1,11 @@
+# utils-generators
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test utils-generators` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Running lint
+
+Run `nx lint utils-generators` to execute the lint via [ESLint](https://eslint.org/).
